autocheck

+++ bluepoint.h
@@ -35,13 +35,13 @@ typedef  unsigned long ulong;
 
 void	bluepoint_encrypt(char *buff, int blen, char *pass, int plen);
 void	bluepoint_decrypt(char *str, int slen, char *pass, int plen);
-ulong   bluepoint_hash(char *buff, int blen);
-ulong   bluepoint_crypthash(char *buff, int blen, char *pass, int plen);
+uint    bluepoint_hash(char *buff, int blen);
+uint    bluepoint_crypthash(char *buff, int blen, char *pass, int plen);
 
 char    *bluepoint_fromhex(char *str, int len, char *out, int *olen);
 char    *bluepoint_tohex(char *str, int len, char *out, int *olen);
 
-unsigned long long   bluepoint_hash64(const char *buff, int blen);
+unsigned long long bluepoint_hash64(const char *buff, int blen);
 unsigned long long bluepoint_crypthash64(const char *buff, int blen, char *pass, int plen);
 
 #ifdef DEF_DUMPHEX
